requirements/base.txt
Django==1.6.6
Pillow==2.5.3
django-bootstrap-form==3.1
django-debug-toolbar==1.2.1
django-guardian==1.2.4
django-userena==1.3.0
easy-thumbnails==2.1
html2text==2014.7.3
six==1.7.3
sqlparse==0.1.11
wsgiref==0.1.2
